move files around, add hush command back
