+++ src/evalCommands.ts
@@ -1,0 +1,22 @@
+import { send } from "./getRepl";
+import { getExpressionUnderCursor } from "./getExpressionUnderCursor";
+
+export const evalCommand = () => {
+  const input = getExpressionUnderCursor(false);
+  if (!input) {
+    return;
+  }
+  send(input);
+}
+
+export const evalMultiCommand =  () => {
+  const input = getExpressionUnderCursor(true);
+  if (!input) {
+      return;
+  }
+  send(input);
+};
+
+export const hushCommand = () => {
+  send('hush');
+}
