--- test/ghci.test.ts
@@ -1,3 +1,0 @@
-suite("GHCi", () => {
-    // TODO: figure out how to write tests against GHCi
-});
